Add argument to skip resetting Windows password
Resetting a password makes changes in the metadata, which forces
the instance to restart. This takes approximately 4 minutes.
To avoid this toll every time the build starts, this change
adds a new flag to skip this process, if we already use a
password we know about (eg. it's hardcoded to the image).